perf.c: fix delta time calculation.
If GET_CPU_CYCLES() is only "unsigned" instead of "unsigned long",
differences should be calculated with "unsigned" precision only to
avoid huge numbers in case of the timer overflow.

+++ lib/pscom/perf.c
@@ -94,8 +94,8 @@ void perf_print(void)
 	if (!cur->id) break;
 	while (1) {
 	    printf("pid_%06d %12.2f %12.2f %2u %20s %lu\n", pid,
-		   (unsigned long)(cur->time - firsttime) * cycles_us,
-		   (unsigned long)(cur->time - lasttime) * cycles_us,
+		   (unsigned /*long*/)(cur->time - firsttime) * cycles_us,
+		   (unsigned /*long*/)(cur->time - lasttime) * cycles_us,
 		   get_id_index(cur->id),
 		   cur->id,
 		   cur->time);
